+++ packages/apollo-angular-link-http-batch/src/types.ts
@@ -1,8 +1,0 @@
-import {Operation} from 'apollo-link';
-import {Options} from 'apollo-angular-link-http-common';
-
-export type BatchOptions = {
-  batchMax?: number;
-  batchInterval?: number;
-  batchKey?: (operation: Operation) => string;
-} & Options;