+++ packages/apollo-angular-link-http/src/index.ts
@@ -1,3 +1,0 @@
-export {HttpLink, HttpLinkHandler} from './http-link';
-export {HttpLinkModule, PROVIDERS} from './http-link-module';
-export {Options} from 'apollo-angular-link-http-common';