+++ packages/apollo-angular/schematics/install/schema.ts
@@ -1,4 +1,6 @@
 export interface Schema {
   /** Name of the project to target. */
   project?: string;
+  /** Url to your GraphQL endpoint */
+  endpoint?: string;
 }