+++ packages/apollo-angular/src/index.ts
@@ -3,7 +3,12 @@ export {QueryRef} from './query-ref';
 export {Query} from './query';
 export {Mutation} from './mutation';
 export {Subscription} from './subscription';
-export {SelectPipe} from './select-pipe';
-export {ApolloModule} from './apollo-module';
-export {APOLLO_OPTIONS, APOLLO_NAMED_OPTIONS} from './tokens';
-export {SubscriptionResult} from './types';
+export {APOLLO_OPTIONS, APOLLO_NAMED_OPTIONS, APOLLO_FLAGS} from './tokens';
+export {
+  NamedOptions,
+  SubscriptionResult,
+  WatchQueryOptions,
+  ExtraSubscriptionOptions,
+  Flags,
+} from './types';
+export {gql} from '@apollo/client/core';